# [Spike] Replace type intersections with interface extensions