--- a/apps/web/src/pages/Communities/CommunityMembersPage/components/EditCommunityMemberDialog.tsx
+++ b/apps/web/src/pages/Communities/CommunityMembersPage/components/EditCommunityMemberDialog.tsx
@@ -32,13 +32,12 @@ import { getTeamBasedRoleOptions } from "./utils";
 import useAvailableRoles from "./useAvailableRoles";
 import { UpdateUserCommunityRoles_Mutation } from "./operations";
 
-type EditCommunityMemberDialogProps = ComponentPropsWithoutRef<
-  typeof DropdownMenu.Item
-> & {
+interface EditCommunityMemberDialogProps
+  extends ComponentPropsWithoutRef<typeof DropdownMenu.Item> {
   user: CommunityMember;
   community: CommunityMembersPageCommunityFragmentType;
   hasPlatformAdmin: boolean;
-};
+}
 
 const EditCommunityMemberDialog = forwardRef<
   ComponentRef<typeof DropdownMenu.Item>,