--- a/apps/web/src/pages/EmployeeProfile/components/CareerDevelopmentSection/CareerDevelopmentSection.tsx
+++ b/apps/web/src/pages/EmployeeProfile/components/CareerDevelopmentSection/CareerDevelopmentSection.tsx
@@ -111,10 +111,11 @@ const dataToFormValues = ({
   execCoachingInterest: execCoachingInterest?.map((x) => x.value) ?? [],
 });
 
-export type FormValues = Pick<
-  UpdateEmployeeProfileInput,
-  "mentorshipInterest" | "execCoachingInterest"
-> & {
+export interface FormValues
+  extends Pick<
+    UpdateEmployeeProfileInput,
+    "mentorshipInterest" | "execCoachingInterest"
+  > {
   lateralMoveInterest?: "yes" | "no";
   lateralMoveTimeFrame?: TimeFrame | null;
   lateralMoveOrganizationType?: OrganizationTypeInterest[] | null;
@@ -127,7 +128,7 @@ export type FormValues = Pick<
   mentorshipStatus?: string | null;
   execInterest?: "yes" | "no";
   execCoachingStatus?: string | null;
-};
+}
 
 interface CareerDevelopmentSectionProps {
   employeeProfileQuery: FragmentType<