--- a/apps/web/src/pages/Pools/AssessmentPlanBuilderPage/components/AssessmentDetailsDialog.tsx
+++ b/apps/web/src/pages/Pools/AssessmentPlanBuilderPage/components/AssessmentDetailsDialog.tsx
@@ -169,13 +169,11 @@ interface FormValues {
   assessedSkillsScreeningQuestions?: Maybe<Scalars["ID"]["output"][]>;
 }
 
-type InitialValues = Omit<
-  FormValues,
-  "poolId" | "screeningQuestionFieldArray"
-> & {
+interface InitialValues
+  extends Omit<FormValues, "poolId" | "screeningQuestionFieldArray"> {
   poolId: NonNullable<FormValues["poolId"]>;
   screeningQuestions?: ScreeningQuestion[];
-};
+}
 
 interface AssessmentDetailsDialogProps {
   initialValues: InitialValues;