--- a/apps/web/src/pages/Pools/EditPoolPage/components/PoolNameSection/PoolNameSection.tsx
+++ b/apps/web/src/pages/Pools/EditPoolPage/components/PoolNameSection/PoolNameSection.tsx
@@ -232,13 +232,14 @@ const allSelectionLimitations: Record<
   },
 } as const;
 
-type PoolNameSectionProps = SectionProps<
-  PoolNameSubmitData,
-  FragmentType<typeof EditPoolName_Fragment>
-> & {
+interface PoolNameSectionProps
+  extends SectionProps<
+    PoolNameSubmitData,
+    FragmentType<typeof EditPoolName_Fragment>
+  > {
   classificationsQuery: FragmentType<typeof PoolClassification_Fragment>[];
   departmentsQuery: FragmentType<typeof PoolDepartment_Fragment>[];
-};
+}
 
 const PoolNameSection = ({
   poolQuery,