🤖 Resolves #14490 

## 👋 Introduction

Replaces our type intersections with interfaces to attempt to improve the performance of typescript

## 🕵️ Details

### TSC

Overall this lead to no difference for `tsc`, the timings were all neglible and can be described by slight differences in repeat runs.

| Metric               | Before (s) | After (s) | Diff (+/-%) |
|----------------------|------------|-----------|-------------|
| I/O Read             | 0.1        | 0.09      | -10.5       |
| Parse                | 1.42       | 1.41      | -0.7        |
| ResolveModule        | 0.3        | 0.32      | +6.45       |
| ResolveTypeReference | 0          | 0         | 0           |
| ResolveLibrary       | 0          | 0         | 0           |
| Program              | 2.05       | 2.05      | 0           |
| Bind                 | 0.87       | 0.99      | +12.9       |
| Check                | 13.4       | 13.27     | -0.97       |
| printTime            | 0          | 0         | 0           |
| Emit                 | 0          | 0         | 0           |
| Total                | 16.31      | 16.31     | 0           |

### ESLint

Of course, eslint uses typescript for its type-aware rules fomr `@typescript-eslint` we have, there was also no noticeable improvement there but not a massive one. Here is a comparisson of the top longest running rules

#### Before

**Total**: 47.109s

Rule                                                   | Time (ms) | Relative
:------------------------------------------------------|----------:|--------:
@typescript-eslint/no-deprecated                       |  9966.917 |    31.6%
@typescript-eslint/no-unsafe-assignment                |  3685.007 |    11.7%
import/default                                         |  3001.021 |     9.5%
turbo/no-undeclared-env-vars                           |  1971.289 |     6.3%
@typescript-eslint/no-misused-promises                 |  1859.049 |     5.9%
@typescript-eslint/no-unnecessary-type-assertion       |  1296.225 |     4.1%
import/no-unresolved                                   |   842.110 |     2.7%
@typescript-eslint/no-unused-vars                      |   744.275 |     2.4%
react/no-direct-mutation-state                         |   632.322 |     2.0%
import/no-named-as-default-member                      |   572.948 |     1.8%

#### After

**Total**: 47.019s

Rule                                                   | Time (ms) | Relative
:------------------------------------------------------|----------:|--------:
@typescript-eslint/no-deprecated                       |  9802.763 |    31.7%
@typescript-eslint/no-unsafe-assignment                |  3671.871 |    11.9%
import/default                                         |  2856.278 |     9.2%
turbo/no-undeclared-env-vars                           |  1997.645 |     6.5%
@typescript-eslint/no-misused-promises                 |  1839.457 |     5.9%
@typescript-eslint/no-unnecessary-type-assertion       |  1221.046 |     3.9%
import/no-unresolved                                   |   848.506 |     2.7%
@typescript-eslint/no-unused-vars                      |   753.569 |     2.4%
react/no-direct-mutation-state                         |   613.393 |     2.0%
import/no-named-as-default-member                      |   555.917 |     1.8%

## :thought_balloon: Thoughts

Even though this doesn't seem to add much performance, it could still be worth merging for the following reasons:

 - Consistency, we were mixing the two methods and we do have a rule to prefer itnerfaces so this aligns with that as well
 - Best practices, since this is usually reccomended and will flag conflicts rather than merging them, I think this method makes sense
 
 > Interfaces create a single flat object type that detects property conflicts, which are usually important to resolve

REF: https://github.com/microsoft/Typescript/wiki/Performance#preferring-interfaces-over-intersections

## 🧪 Testing

1. Run `pnpm tsc -- --extendedDiagnostics`
2. Confirm it passes in a reasonable time (whatever that means to you)
3. Confirm no errors